Standalone MiniJson.cs
