+++ Deroute/DerouteSharp.csproj
@@ -73,6 +73,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CellSupport.cs" />
+    <Compile Include="Collab\MiniJson.cs" />
     <Compile Include="FormAbout.cs">
       <SubType>Form</SubType>
     </Compile>
